Changed longs to doubles for lng / lat, implemented test case for one API call.

+++ MarksMenusClient/src/com/marksmenus/controller/APIMap.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 
 public interface APIMap {
 
-	public ArrayList<Restaurant> findRestuarantsByLocation(long lat, long lng, int distance);
+	public ArrayList<Restaurant> findRestuarantsByLocation(double lat, double lng, int distance);
 	public ArrayList<Restaurant> findRestaurantsByKeyword(long lat, long lng, int distance, String keyword);
 	public Restaurant getRestaurant(int id);
 	public Menu getMenu(int id);
